refactor: userRepository data Layer 의존성 제거
