+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetUserBoardPagerUseCase.kt
@@ -1,91 +1,20 @@
 package com.smilehunter.ablebody.domain.usecase
 
-import androidx.paging.Pager
-import androidx.paging.PagingConfig
 import androidx.paging.PagingData
-import androidx.paging.PagingSource
-import androidx.paging.PagingState
 import com.smilehunter.ablebody.domain.model.UserBoardData
 import com.smilehunter.ablebody.domain.repository.UserRepository
-import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
-import com.smilehunter.ablebody.network.di.Dispatcher
-import com.smilehunter.ablebody.network.model.response.GetMyBoardResponseData
-import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class GetUserBoardPagerUseCase @Inject constructor(
-    @Dispatcher(AbleBodyDispatcher.IO) private val ioDispatcher: CoroutineDispatcher,
     private val userRepository: UserRepository
 ) {
 
     operator fun invoke(): Flow<PagingData<UserBoardData.Content>> {
-        return Pager(
-            config = PagingConfig(pageSize = 10),
-            initialKey = 0
-        ) {
-            UserBoardPagingSource()
-        }
-            .flow
+        return userRepository.getMyBoard(null)
     }
 
     operator fun invoke(uid: String): Flow<PagingData<UserBoardData.Content>> {
-        return Pager(
-            config = PagingConfig(pageSize = 10),
-            initialKey = 0
-        ) {
-            UserBoardPagingSource(uid = uid)
-        }
-            .flow
-    }
-
-    inner class UserBoardPagingSource(
-        private val uid: String? = null
-    ): PagingSource<Int, UserBoardData.Content>() {
-        override fun getRefreshKey(state: PagingState<Int, UserBoardData.Content>): Int? {
-            return state.anchorPosition
-        }
-
-        override suspend fun load(params: LoadParams<Int>): LoadResult<Int, UserBoardData.Content> =
-            try {
-                val currentPageIndex = params.key ?: 0
-
-                val userBoardData = withContext(ioDispatcher) {
-                    userRepository.getMyBoard(
-                        uid = uid,
-                        page = currentPageIndex,
-                        size = params.loadSize
-                    ).data!!.toDomain()
-                }
-
-                LoadResult.Page(
-                    data = userBoardData.content,
-                    prevKey = if (userBoardData.first) null else userBoardData.number - 1,
-                    nextKey = if (userBoardData.last) null else userBoardData.number + 1
-                )
-
-            } catch (e: Exception) {
-                LoadResult.Error(e)
-            }
-
+        return userRepository.getMyBoard(uid)
     }
 }
-
-private fun GetMyBoardResponseData.toDomain() = UserBoardData(
-    content = content.map {
-        UserBoardData.Content(
-            id = it.id,
-            imageURL = it.imageURL,
-            createDate = it.createDate,
-            comments = it.comments,
-            likes = it.likes,
-            views = it.views,
-            isSingleImage = it.plural,
-        )
-    },
-    totalPages = totalPages,
-    last = last,
-    number = number,
-    first = first
-)
